751 simplify and redeploy react hooks docs (#752)
* Replace text generation card with TextTab component

- Simplify text generation UI by using TextTab component
- Remove redundant getTextCode function
- Enhance code readability and maintainability

* Create TextTab.tsx

* Remove unused imports and code for text and image generation

- Removed usePollinationsText and usePollinationsImage imports
- Deleted state variables related to text and image prompts
- Removed text and image generation logic
- Updated ImageTab component usage
- Cleaned up unused functions and code blocks

* Remove unused imports and chat-related code

- Removed KeyboardEvent import
- Removed Send import
- Removed chat state and handlers
- Removed chat UI components
- Integrated ChatTab component

* create chattab and imagetab components

* extract the model fetching logic into a separate hook

* Update component props and state management

- Remove unused props from ChatTab, ImageTab, and TextTab
- Integrate useFetchModels hook directly within components
- Simplify state initialization for selected models
- Clean up PollinationsDemo component by removing redundant state and props

* Remove unused imports from page.tsx

- Removed Card, CardContent, CardDescription, CardHeader, CardTitle
- Removed Label, Input, Button, Textarea
- Removed Select, SelectContent, SelectItem, SelectTrigger, SelectValue
- Removed Copy from 'lucide-react'
- Removed ReactMarkdown and SyntaxHighlighter
- Removed oneDark from 'react-syntax-highlighter/dist/esm/styles/prism'

* Add debounced image fetching

- Integrate useDebounce hook for image fetching
- Replace usePollinationsImage with debounced version
- Improve performance by reducing API calls

* Add debouncing to text prompt

- Implement usePollinationsTextDebounced to debounce text prompt and options
- Replace usePollinationsText with usePollinationsTextDebounced in TextTab component
- Improve performance by reducing the number of API calls

* Enhance hooks UI and layout

- Make system message input full-width
- Place seed and model fields on the same line
- Highlight send button with hover effect
- Improve responsive design for various screen sizes
- Refactor component for better code organization
- Enhance accessibility with proper labeling and semantic structure

* Enhance hooks UI and layout

- Make system message input full-width
- Place seed and model fields on the same line
- Highlight send button with hover effect
- Improve responsive design for various screen sizes
- Refactor component for better code organization
- Enhance accessibility with proper labeling and semantic structure

* Update container class with max-width

- Added max-w-4xl to container div for better layout control
- Ensures content does not stretch too wide on larger screens

* Simplify tab rendering logic

- Create a tabs array to store tab data
- Use map function to render TabsTrigger and TabsContent
- Remove redundant code for individual tab rendering

* Remove unused refs and keypress handlers

- Removed useRef and useEffect hooks for chat container scrolling
- Removed handleKeyPress function for sending messages on Enter key
- Updated handleSendMessage to handle button click event
- Cleaned up related code in getChatCode function

* Update package version and add TypeScript types

- Bump version from 2.0.1 to 2.0.2
- Add TypeScript types definition file
- Improve TypeScript support for the package

* Create index.d.ts

* bump pollinations react version that hopefully includes types

* Delete legacy page

* update @pollinations/react to incldue type definitions

* Update README.md

* Create netlify.toml

* Update components to use concise function syntax

- Remove React.FC type annotations
- Simplify component declarations
- Ensure consistency across all components

* Move netlify.toml to app folder and update @pollinations/react package

* Update README links and fix GitHub URL in PollinationsDemo

- Update React hooks link in main README
- Add link to interactive examples in pollinations-react README
- Correct GitHub URL in PollinationsDemo component

* Update Netlify config for Node 18
- Specify node_version as 18 in netlify.toml
- Ensure compatibility with latest Node version

---------
