+++ app/package.json
@@ -9,7 +9,7 @@
     "@material-ui/core": "^4.12.4",
     "@material-ui/icons": "^4.11.2",
     "@mui/material": "^6.1.0",
-    "@pollinations/react": "^2.0.1",
+    "@pollinations/react": "^2.0.3",
     "debug": "^4.3.2",
     "front-matter": "^4.0.2",
     "json5": "^2.2.1",
