+++ pollinations-react/docs/app/page.tsx
@@ -1,244 +1,21 @@
 'use client'
-// @ts-expect-error todo: interfaces
-import { usePollinationsText, usePollinationsImage, usePollinationsChat } from '@pollinations/react'
-import React, { useState, useEffect, KeyboardEvent, useRef } from 'react'
+import React, { useState } from 'react'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import { Label } from "@/components/ui/label"
-import { Input } from "@/components/ui/input"
-import { Button } from "@/components/ui/button"
-import { Textarea } from "@/components/ui/textarea"
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select"
-import { Copy, Send } from 'lucide-react'
-import ReactMarkdown from 'react-markdown'
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter'
-import { oneDark } from 'react-syntax-highlighter/dist/esm/styles/prism'
+import TextTab from './components/TextTab';
+import ImageTab from './components/ImageTab';
+import ChatTab from './components/ChatTab';
 
-interface TextModel {
-  name: string
-  type: 'chat' | 'completion'
-  censored: boolean
-}
-
-type ImageModel = string
+const tabs = [
+  { value: 'text', label: 'usePollinationsText', shortLabel: 'Text', component: <TextTab /> },
+  { value: 'image', label: 'usePollinationsImage', shortLabel: 'Image', component: <ImageTab /> },
+  { value: 'chat', label: 'usePollinationsChat', shortLabel: 'Chat', component: <ChatTab /> }
+];
 
 export default function PollinationsDemo() {
-  const [activeTab, setActiveTab] = useState<'text' | 'image' | 'chat'>('text')
-  const [textPrompt, setTextPrompt] = useState("Write a haiku about artificial intelligence")
-  const [imagePrompt, setImagePrompt] = useState("A futuristic city with flying cars and neon lights")
-  const [chatPrompt, setChatPrompt] = useState("")
-  const [textModels, setTextModels] = useState<TextModel[]>([])
-  const [imageModels, setImageModels] = useState<ImageModel[]>([])
-  const [selectedTextModel, setSelectedTextModel] = useState<string>('openai')
-  const [selectedImageModel, setSelectedImageModel] = useState<string>('flux')
-  const [textSeed, setTextSeed] = useState<number>(42)
-  const [imageSeed, setImageSeed] = useState<number>(42)
-  const [imageWidth, setImageWidth] = useState<number>(1024)
-  const [imageHeight, setImageHeight] = useState<number>(1024)
-  const [systemMessage, setSystemMessage] = useState<string>("You are a helpful AI assistant.")
-  const [chatSeed, setChatSeed] = useState<number>(42)
-  const [chatModel, setChatModel] = useState<string>('openai')
-
-  const chatContainerRef = useRef<HTMLDivElement>(null)
-
-  const { sendUserMessage, messages } = usePollinationsChat([
-    { role: "system", content: systemMessage }
-  ], {
-    seed: chatSeed,
-    model: chatModel
-  })
-
-  useEffect(() => {
-    const fetchModels = async () => {
-      try {
-        const [textResponse, imageResponse] = await Promise.all([
-          fetch('https://text.pollinations.ai/models'),
-          fetch('https://image.pollinations.ai/models')
-        ])
-        const textData: TextModel[] = await textResponse.json()
-        const imageData: ImageModel[] = await imageResponse.json()
-        setTextModels(textData)
-        setImageModels(imageData)
-      } catch (error) {
-        console.error('Error fetching models:', error)
-        // Fallback models
-        setTextModels([{ name: 'openai', type: 'chat', censored: false }])
-        setImageModels(['flux'])
-      }
-    }
-    fetchModels()
-  }, [])
-
-  useEffect(() => {
-    if (chatContainerRef.current) {
-      chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight
-    }
-  }, [messages])
-
-  const textResult = usePollinationsText(textPrompt, {
-    seed: textSeed,
-    model: selectedTextModel
-  })
-
-  const imageUrl = usePollinationsImage(imagePrompt, {
-    width: imageWidth,
-    height: imageHeight,
-    seed: imageSeed,
-    model: selectedImageModel,
-    nologo: true
-  })
-
-  const handleSendMessage = () => {
-    if (chatPrompt.trim()) {
-      sendUserMessage(chatPrompt)
-      setChatPrompt('')
-    }
-  }
-
-  const handleKeyPress = (event: KeyboardEvent<HTMLTextAreaElement>) => {
-    if (event.key === 'Enter' && !event.shiftKey) {
-      event.preventDefault()
-      handleSendMessage()
-    }
-  }
-
-  const getTextCode = (): string => {
-    return `
-import React from 'react';
-import { usePollinationsText } from '@pollinations/react';
-import ReactMarkdown from 'react-markdown';
-
-const TextComponent: React.FC = () => {
-  const text = usePollinationsText("${textPrompt}", { 
-    seed: ${textSeed},
-    model: '${selectedTextModel}'
-  });
-  
-  return (
-    <div>
-      {text ? <ReactMarkdown>{text}</ReactMarkdown> : <p>Loading...</p>}
-    </div>
-  );
-};
-
-export default TextComponent;
-    `
-  }
-
-  const getImageCode = (): string => {
-    const imageUrlWithParams = `https://image.pollinations.ai/prompt/${encodeURIComponent(imagePrompt)}?width=${imageWidth}&height=${imageHeight}&seed=${imageSeed}&model=${selectedImageModel}&nologo=true`
-
-    return `
-import React from 'react';
-import { usePollinationsImage } from '@pollinations/react';
-
-const ImageComponent: React.FC = () => {
-  const imageUrl = usePollinationsImage("${imagePrompt}", {
-    width: ${imageWidth},
-    height: ${imageHeight},
-    seed: ${imageSeed},
-    model: '${selectedImageModel}',
-    nologo: true
-  });
+  const [activeTab, setActiveTab] = useState<'text' | 'image' | 'chat'>('text');
 
   return (
-    <div>
-      {imageUrl ? (
-        <img 
-          src={imageUrl} 
-          alt="Generated image" 
-          style={{
-            width: '100%',
-            height: '100%',
-            objectFit: 'contain',
-          }}
-        />
-      ) : (
-        <p>Loading...</p>
-      )}
-    </div>
-  );
-};
-
-export default ImageComponent;
-    `
-  }
-
-  const getChatCode = (): string => {
-    return `
-import React, { useState, useRef, useEffect } from 'react';
-import { usePollinationsChat } from '@pollinations/react';
-import ReactMarkdown from 'react-markdown';
-
-const ChatComponent: React.FC = () => {
-  const [input, setInput] = useState('');
-  const chatContainerRef = useRef<HTMLDivElement>(null);
-  const { sendUserMessage, messages } = usePollinationsChat([
-    { role: "system", content: "${systemMessage}" }
-  ], { 
-    seed: ${chatSeed},
-    model: '${chatModel}'
-  });
-
-  useEffect(() => {
-    if (chatContainerRef.current) {
-      chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight;
-    }
-  }, [messages]);
-
-  const handleSend = () => {
-    if (input.trim()) {
-      sendUserMessage(input);
-      setInput('');
-    }
-  };
-
-  const handleKeyPress = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    if (event.key === 'Enter' && !event.shiftKey) {
-      event.preventDefault();
-      handleSend();
-    }
-  };
-
-  return (
-    <div className="flex flex-col h-[500px]">
-      <div ref={chatContainerRef} className="flex-1 overflow-y-auto p-4 space-y-4">
-        {messages.map((msg, index) => (
-          <div key={index} className={\`flex \${msg.role === 'user' ? 'justify-end' : 'justify-start'}\`}>
-            <div className={\`max-w-[70%] p-3 rounded-lg \${
-              msg.role === 'user' ? 'bg-blue-500 text-white' : 'bg-gray-200 text-black'
-            }\`}>
-              {msg.role === 'user' ? '🐦' : '🌸'} <ReactMarkdown>{msg.content}</ReactMarkdown>
-            </div>
-          </div>
-        ))}
-      </div>
-      <div className="p-4 border-t">
-        <input
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          onKeyPress={handleKeyPress}
-          placeholder="Type your message..."
-          className="w-full p-2 border rounded-lg"
-        />
-        <button onClick={handleSend} className="mt-2 px-4 py-2 bg-blue-500 text-white rounded-lg">
-          Send
-        </button>
-      </div>
-    </div>
-  );
-};
-
-export default ChatComponent;
-    `
-  }
-
-  const copyToClipboard = (text: string) => {
-    navigator.clipboard.writeText(text)
-  }
-
-  return (
-    <div className="container mx-auto p-4 bg-slate-900 text-slate-100">
+    <div className="container mx-auto p-4 bg-slate-900 text-slate-100 max-w-4xl">
       <h1 className="text-4xl font-bold mb-2 text-center">
         <a href="https://github.com/pollinations/pollinations" target="_blank" rel="noopener noreferrer" className="hover:text-blue-400 transition-colors">
           🌸 Pollinations Generative React Hooks 2.0.1 🌸
@@ -248,286 +25,22 @@ export default ChatComponent;
 
       <Tabs value={activeTab} onValueChange={(value) => setActiveTab(value as 'text' | 'image' | 'chat')} className="w-full">
         <TabsList className="grid w-full grid-cols-3">
-          <TabsTrigger
-            value="text"
-            className={`${activeTab === 'text' ? 'bg-slate-300 text-white underline' : 'bg-slate-800 text-slate-500'} transition-colors duration-200 hover:bg-slate-600`}
-          >
-            <span className="hidden sm:inline">usePollinationsText</span>
-            <span className="sm:hidden">Text</span>
-          </TabsTrigger>
-          <TabsTrigger
-            value="image"
-            className={`${activeTab === 'image' ? 'bg-slate-300 text-white underline' : 'bg-slate-800 text-slate-500'} transition-colors duration-200 hover:bg-slate-600`}
-          >
-            <span className="hidden sm:inline">usePollinationsImage</span>
-            <span className="sm:hidden">Image</span>
-          </TabsTrigger>
-          <TabsTrigger
-            value="chat"
-            className={`${activeTab === 'chat' ? 'bg-slate-300 text-white underline' : 'bg-slate-800 text-slate-500'} transition-colors duration-200 hover:bg-slate-600`}
-          >
-            <span className="hidden sm:inline">usePollinationsChat</span>
-            <span className="sm:hidden">Chat</span>
-          </TabsTrigger>
+          {tabs.map((tab) => (
+            <TabsTrigger
+              key={tab.value}
+              value={tab.value}
+              className={`${activeTab === tab.value ? 'bg-slate-300 text-white underline' : 'bg-slate-800 text-slate-500'} transition-colors duration-200 hover:bg-slate-600`}
+            >
+              <span className="hidden sm:inline">{tab.label}</span>
+              <span className="sm:hidden">{tab.shortLabel}</span>
+            </TabsTrigger>
+          ))}
         </TabsList>
-        <TabsContent value="text">
-          <Card className="bg-slate-800 text-slate-100">
-            <CardHeader>
-              <CardTitle>Text Generation</CardTitle>
-              <CardDescription>Generate text using Pollinations' API</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex flex-col space-y-4">
-                <div>
-                  <Label htmlFor="textPrompt">Prompt</Label>
-                  <Input
-                    id="textPrompt"
-                    value={textPrompt}
-                    onChange={(e) => setTextPrompt(e.target.value)}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="textModel">Model</Label>
-                  <Select
-                    value={selectedTextModel}
-                    onValueChange={setSelectedTextModel}
-                  >
-                    <SelectTrigger id="textModel" className="bg-slate-700 text-slate-100">
-                      <SelectValue placeholder="Select a model" />
-                    </SelectTrigger>
-                    <SelectContent className="bg-slate-700 text-slate-100">
-                      {textModels.map((model) => (
-                        <SelectItem key={model.name} value={model.name}>{model.name}</SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-                <div>
-                  <Label htmlFor="textSeed">Seed</Label>
-                  <Input
-                    id="textSeed"
-                    type="number"
-                    value={textSeed}
-                    onChange={(e) => setTextSeed(Math.max(1, Number(e.target.value)))}
-                    min={1}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <h3 className="text-lg font-semibold mb-2">Generated Text:</h3>
-                  <div className="bg-slate-700 p-4 rounded-md">
-                    {textResult ? (
-                      <ReactMarkdown>{textResult}</ReactMarkdown>
-                    ) : (
-                      <p>Loading...</p>
-                    )}
-                  </div>
-                </div>
-                <div>
-                  <h3 className="text-lg font-semibold mb-2">Code Preview:</h3>
-                  <div className="relative">
-                    <SyntaxHighlighter language="typescript" style={oneDark} className="rounded-md">
-                      {getTextCode()}
-                    </SyntaxHighlighter>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="absolute top-2 right-2"
-                      onClick={() => copyToClipboard(getTextCode())}
-                    >
-                      <Copy className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-        <TabsContent value="image">
-          <Card className="bg-slate-800 text-slate-100">
-            <CardHeader>
-              <CardTitle>Image Generation</CardTitle>
-              <CardDescription>Generate images using Pollinations' API</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex flex-col space-y-4">
-                <div>
-                  <Label htmlFor="imagePrompt">Prompt</Label>
-                  <Input
-                    id="imagePrompt"
-                    value={imagePrompt}
-                    onChange={(e) => setImagePrompt(e.target.value)}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="imageModel">Model</Label>
-                  <Select
-                    value={selectedImageModel}
-                    onValueChange={setSelectedImageModel}
-                  >
-                    <SelectTrigger id="imageModel" className="bg-slate-700 text-slate-100">
-                      <SelectValue placeholder="Select a model" />
-                    </SelectTrigger>
-                    <SelectContent className="bg-slate-700 text-slate-100">
-                      {imageModels.map((model) => (
-                        <SelectItem key={model} value={model}>{model}</SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-                <div>
-                  <Label htmlFor="imageSeed">Seed</Label>
-                  <Input
-                    id="imageSeed"
-                    type="number"
-                    value={imageSeed}
-                    onChange={(e) => setImageSeed(Math.max(1, Number(e.target.value)))}
-                    min={1}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="imageWidth">Width</Label>
-                  <Input
-                    id="imageWidth"
-                    type="number"
-                    value={imageWidth}
-                    onChange={(e) => setImageWidth(Math.max(32, Number(e.target.value)))}
-                    min={32}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="imageHeight">Height</Label>
-                  <Input
-                    id="imageHeight"
-                    type="number"
-                    value={imageHeight}
-                    onChange={(e) => setImageHeight(Math.max(32, Number(e.target.value)))}
-                    min={32}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <h3 className="text-lg font-semibold mb-2">Generated Image:</h3>
-                  {imageUrl ? (
-                    <img
-                      src={imageUrl}
-                      alt="Generated image"
-                      className="w-full h-auto rounded-md"
-                    />
-                  ) : (
-                    <p>Loading...</p>
-                  )}
-                </div>
-                <div>
-                  <h3 className="text-lg font-semibold mb-2">Code Preview:</h3>
-                  <div className="relative">
-                    <SyntaxHighlighter language="typescript" style={oneDark} className="rounded-md">
-                      {getImageCode()}
-                    </SyntaxHighlighter>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="absolute top-2 right-2"
-                      onClick={() => copyToClipboard(getImageCode())}
-                    >
-                      <Copy className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-        <TabsContent value="chat">
-          <Card className="bg-slate-800 text-slate-100">
-            <CardHeader>
-              <CardTitle>Chat</CardTitle>
-              <CardDescription>Chat with Pollinations' AI</CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="flex flex-col space-y-4">
-                <div>
-                  <Label htmlFor="systemMessage">System Message</Label>
-                  <Input
-                    id="systemMessage"
-                    value={systemMessage}
-                    onChange={(e) => setSystemMessage(e.target.value)}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="chatSeed">Seed</Label>
-                  <Input
-                    id="chatSeed"
-                    type="number"
-                    value={chatSeed}
-                    onChange={(e) => setChatSeed(Math.max(1, Number(e.target.value)))}
-                    min={1}
-                    className="bg-slate-700 text-slate-100"
-                  />
-                </div>
-                <div>
-                  <Label htmlFor="chatModel">Model</Label>
-                  <Select
-                    value={chatModel}
-                    onValueChange={setChatModel}
-                  >
-                    <SelectTrigger id="chatModel" className="bg-slate-700 text-slate-100">
-                      <SelectValue placeholder="Select a model" />
-                    </SelectTrigger>
-                    <SelectContent className="bg-slate-700 text-slate-100">
-                      {textModels.map((model) => (
-                        <SelectItem key={model.name} value={model.name}>{model.name}</SelectItem>
-                      ))}
-                    </SelectContent>
-                  </Select>
-                </div>
-                <div ref={chatContainerRef} className="h-64 overflow-y-auto bg-slate-700 p-4 rounded-md space-y-4">
-                  {messages.map((msg: any, index: number) => (
-                    <div key={index} className={`flex ${msg.role === 'user' ? 'justify-end' : 'justify-start'}`}>
-                      <div className={`max-w-[70%] p-3 rounded-lg ${msg.role === 'user' ? 'bg-blue-500 text-white' : 'bg-gray-200 text-black'
-                        }`}>
-                        {msg.role === 'user' ? '🐦' : '🌸'} <ReactMarkdown>{msg.content}</ReactMarkdown>
-                      </div>
-                    </div>
-                  ))}
-                </div>
-                <div className="flex space-x-2">
-                  <Textarea
-                    value={chatPrompt}
-                    onChange={(e) => setChatPrompt(e.target.value)}
-                    onKeyPress={handleKeyPress}
-                    placeholder="Type your message..."
-                    className="bg-slate-700 text-slate-100 flex-grow"
-                  />
-                  <Button onClick={handleSendMessage}>
-                    <Send className="h-4 w-4" />
-                  </Button>
-                </div>
-                <div>
-                  <h3 className="text-lg font-semibold mb-2">Code Preview:</h3>
-                  <div className="relative">
-                    <SyntaxHighlighter language="typescript" style={oneDark} className="rounded-md">
-                      {getChatCode()}
-                    </SyntaxHighlighter>
-                    <Button
-                      variant="ghost"
-                      size="sm"
-                      className="absolute top-2 right-2"
-                      onClick={() => copyToClipboard(getChatCode())}
-                    >
-                      <Copy className="h-4 w-4" />
-                    </Button>
-                  </div>
-                </div>
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
+        {tabs.map((tab) => (
+          <TabsContent key={tab.value} value={tab.value}>
+            {tab.component}
+          </TabsContent>
+        ))} 
       </Tabs>
 
       <footer className="mt-8 text-center text-slate-300">
@@ -535,7 +48,7 @@ export default ChatComponent;
          &nbsp;and&nbsp;
           <a className='hover:underline' href="https://karma.yt" title='Karma.yt'>Karma.yt</a></p>
         <div className="mt-2 flex justify-center space-x-4">
-          <a href="https://github.com/pollinations/pollinations" target="_blank" rel="noopener noreferrer" className="text-blue-400 hover:text-blue-300">
+          <a href="https://github.com/pollinations/pollinations/pollinations-react" target="_blank" rel="noopener noreferrer" className="text-blue-400 hover:text-blue-300">
             View on GitHub
           </a>
           <a href="https://www.npmjs.com/package/@pollinations/react" target="_blank" rel="noopener noreferrer" className="text-blue-400 hover:text-blue-300">
