+++ pollinations-react/docs/netlify.toml
@@ -1,0 +1,12 @@
+[build]
+  command = "npm install && npm run build"
+  publish = ".next"  # This path is relative to pollinations-react/docs
+  node_version = "18"
+
+[[plugins]]
+  package = "@netlify/plugin-nextjs"
+
+[[redirects]]
+  from = "/*"
+  to = "/"
+  status = 200
